backend/requirements.txt
# Add Python dependencies here

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pinecone-client==2.2.4
pyairtable==2.1.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
wave==0.0.2
pybase64
datetime==5.3
uuid==1.30
asyncio==3.4.3